End To End QNA chatbot/requirements.txt
langchain_community
langchain_groq
python-dotenv
langchain
langchain_core
streamlit